Update web-scraperFINAL.py
I added a contingency for Beautiful Soup just in case it tries to fail again. I had this problem a lot with trying to get my program to run and this block of code I added helped ensure that whoever used my program didn't just encounter a large pile of errors from not being able to run Beautiful Soup.